server/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-admin-interface==0.24.2
django-autoslug==1.9.8
django-ckeditor-5==0.2.3
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
mysqlclient==2.1.1
Pillow==9.4.0
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
